[FIX] 메니저에서 홈으로 이동안되는 문제 해결

+++ LOTTERIA_Kiosk/LOTTERIA_Kiosk/View/PaymentCompleted.xaml.cs
@@ -100,7 +100,7 @@ namespace LOTTERIA_Kiosk.View
                 OrderMenu orderMenu = new OrderMenu
                 {
                     Name = food.Name,
-                    Price = food.Price / food.Count,
+                    Price = (int)food.Price / food.Count,
                     Count = food.Count
                 };
